Add Rent Photos

+++ kurye_takip/lib/model/get_cars.dart
@@ -3,7 +3,7 @@
 //     final getCars = getCarsFromJson(jsonString);
 
 // ignore_for_file: unnecessary_question_mark, constant_identifier_names
-
+/*
 import 'dart:convert';
 
 GetCars getCarsFromJson(String str) => GetCars.fromJson(json.decode(str));
@@ -45,7 +45,7 @@ class Car {
   int? modelId;
   FuelType? fuelType;
   TransmissionType? transmissionType;
-  int? dailyPrice;
+  double? dailyPrice;
   String? plate;
   Km? km;
   Note? note;
@@ -181,3 +181,4 @@ class EnumValues<T> {
     return reverseMap;
   }
 }
+*/
